+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 All notable changes to this project will be documented in this file.
 (Skipped major version 1, 2 and 3 to match versioning of IDS DataSpaceConnector)
 
-## [8.5.0] - XXXX-XX-XX
+## [8.5.0] - 2021-09-30
 
 ### Added
 - Create, update and delete app stores